requirements.txt
asgiref
certifi
cffi
charset-normalizer
coreapi
coreschema
coverage
cryptography
defusedxml
Django
django-debug-toolbar
django-ranged-response
django-simple-captcha
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
idna
itypes
Jinja2
MarkupSafe
oauthlib
Pillow
psycopg2-binary
pycparser
PyJWT
python3-openid
pytz
requests
requests-oauthlib
six
social-auth-app-django
social-auth-core
sqlparse
uritemplate
urllib3
